i'm an idiot;

+++ src/host/network/Network.cpp
@@ -613,7 +613,7 @@ void Network::clock(uint32_t ms)
                 serverNow = (serverNow << 8) + buffer[13U];
 
                 // check the ping RTT and report any over the maximum defined time
-                uint64_t dt = (uint64_t)fabs(now - serverNow);
+                uint64_t dt = (uint64_t)fabs((double)now - (double)serverNow);
                 if (dt > MAX_SERVER_DIFF)
                     LogWarning(LOG_NET, "PEER %u pong, time delay greater than %llums, now = %llu, server = %llu, dt = %llu", m_peerId, MAX_SERVER_DIFF, now, serverNow, dt);
             }
